an app for locating a nearby shelter during a missile attack
The app locates the user's location and provides him with a route to the nearest shelter. In addition, authorized users can add shelters to the database. The development of this application requires an interface to google maps in order to display a map and route and for calculation of distances and times.
The application was developed in the following languages:
Database: SQL, Server: C#, Client: Angular, PrimeNG.